requirements.txt
asgiref==3.3.4
backcall==0.2.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==5.0.9
Django==3.2.3
django-cors-headers==3.7.0
django-cpf==0.1.0
djangorestframework==3.12.4
google-auth==1.30.1
google-auth-oauthlib==0.4.4
gspread==3.7.0
idna==2.10
ipdb==0.13.7
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
oauthlib==3.1.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
